added casts of iov_len and copyn because iov_len is int on some platforms and size_t on others

+++ ace/SOCK_Dgram.cpp
@@ -251,7 +251,10 @@ ACE_SOCK_Dgram::recv (iovec iov[],
 	   i++)
 	{
 	  ACE_OS::memcpy (iov[i].iov_base, ptr,
-			  copyn > iov[i].iov_len ? iov[i].iov_len : copyn);
+                          // iov_len is int on some platforms, size_t on others
+			  copyn > (int) iov[i].iov_len
+                            ? (size_t) iov[i].iov_len
+                            : (size_t) copyn);
 	  ptr += iov[i].iov_len;
 	  copyn -= iov[i].iov_len;
 	}
